Use top-level example group as example source file
Instead of attributing examples to the file where their immediate parent
example group was defined, use their top-level parent example group.

This means that shared examples are attributed to the files where they
are included, not the files where they are defined. This makes splitting
a test suite based on the runtime of each file more accurate.

The RSpec 2 version already did this; now the RSpec 3 one does too.