Skip STDOUT/STDERR capture when reports are not generated
Using the example `config.around(:each)` behavior listed here has a strange side effect with tools like `pry` (and possibly others) in which invoking `binding.pry` seems to simply be ignored.  This is probably to do with `pry`'s expectation of working with a terminal and finding none.   

But in any case, this code checks whether the `RspecJunitFormatter` has been requested and captures output conditionally on that setting.