Fix back compatiblity
We need to make sure these strings are muable, but `String#+@` is only
available in Ruby 2.3+ and the gem doesn't yet restrict the Ruby version
we're compatible with.

Fix backwards compatiblity by releasing a version using `String#dup`,
then we can add the Ruby version restriction in the Gemfile and revert
to `String#+@`.